+++ src/components/custom/metaballs/index.ts
@@ -1,4 +1,4 @@
 export { default as MetaballCanvas } from "./MetaballCanvas.vue";
-export { useMetaballs } from "./useMetaballs";
+export { useMetaballs, isWebGLSupported } from "./useMetaballs";
 export type { MetaballConfig } from "./types";
 export { DEFAULT_METABALL_CONFIG } from "./types";